+++ R/print.lcmm.R
@@ -7,11 +7,11 @@
 #' 
 #' 
 #' @aliases print.hlme print.lcmm print.Jointlcmm print.multlcmm print.epoce
-#' print.Diffepoce
-#' @param x an object inheriting from classes \code{hlme}, \code{lcmm} for
-#' fitted latent class mixed-effects, or class \code{Jointlcmm}for a Joint
-#' latent class mixed model or \code{epoce} for predictive accuracy
-#' computations.
+#' print.Diffepoce print.externSurv print.externX print.mpjlcmm
+#' @param x an object inheriting from classes \code{hlme}, \code{lcmm}, \code{multlcmm}
+#' for fitted latent class mixed-effects, or class \code{Jointlcmm}, code{mpjclmm} for
+#' a Joint latent class mixed model or \code{epoce} for predictive accuracy
+#' computations or \code{externSurv}, \code{externX} for secondary regression models.
 #' @param \dots further arguments to be passed to or from other methods.  They
 #' are ignored in this function.
 #' @author Cecile Proust-Lima, Viviane Philipps, Amadou Diakite and Benoit
@@ -26,8 +26,17 @@ print.lcmm <-
 function(x,...){
 if (!inherits(x, "lcmm")) stop("use only with \"lcmm\" objects")
 
-cat("General latent class mixed model", "\n")
+if(inherits(x, "externVar")){
+  cat("Secondary linear mixed model", "\n")
+} else {
+  cat("General latent class mixed model", "\n")
+}
 cat("     fitted by maximum likelihood method", "\n")
+if(inherits(x, "externVar")){
+  if(x$varest == "none") cat("     primary model variance not accounted for", "\n")
+  if(x$varest == "Hessian") cat("     primary model variance accounted for through the hessian of the joint likelihood", "\n")
+  if(x$varest == "paramBoot") cat("     primary model variance accounted for through parametric boostrap", "\n")
+}
 
 cl <- x$call
 cl$B <- NULL
@@ -77,9 +86,21 @@ cat("     The program stopped abnormally. No results can be displayed.\n")
 
 cat(" \n")
 cat("     Number of iterations: ", x$niter, "\n")
-cat("     Convergence criteria: parameters=", signif(x$gconv[1],2), "\n")
-cat("                         : likelihood=", signif(x$gconv[2],2), "\n") 
-cat("                         : second derivatives=", signif(x$gconv[3],2), "\n")
+if(inherits(x, "externVar")) {
+  if(x$varest == "paramBoot"){
+    cat("     Proportion of convergence on bootstrap iterations (%)=", x$Mconv, "\n")
+  } else {
+    cat("     Number of iterations: ", x$niter, "\n")
+    cat("     Convergence criteria: parameters=", signif(x$gconv[1],2), "\n")
+    cat("                         : likelihood=", signif(x$gconv[2],2), "\n")
+    cat("                         : second derivatives=", signif(x$gconv[3],2), "\n")
+  }
+} else {
+  cat("     Number of iterations: ", x$niter, "\n")
+  cat("     Convergence criteria: parameters=", signif(x$gconv[1],2), "\n")
+  cat("                         : likelihood=", signif(x$gconv[2],2), "\n")
+  cat("                         : second derivatives=", signif(x$gconv[3],2), "\n")
+}
 cat(" \n")
 cat("Goodness-of-fit statistics:", "\n")
 cat(paste("     maximum log-likelihood:", round(x$loglik,2))," \n")