+++ R/standardMethods.R
@@ -1,7 +1,8 @@
 
 #' Standard methods for estimated models
 #' 
-#' coef, vcov, fixef, ranef, fitted and residuals methods for estimated hlme,
+#' coef and vcov for hlme, lcmm, mutlcmm, Jointlcmm, mpjlcmm, externSurv, and externX
+#' models, fixef, ranef, fitted and residuals methods for estimated hlme,
 #' lcmm, mutlcmm and Jointlcmm models.
 #' 
 #' 
@@ -9,8 +10,9 @@
 #' 
 #'
 #' 
-#' @aliases coef.hlme coef.lcmm coef.Jointlcmm coef.multlcmm vcov.hlme
-#' vcov.lcmm vcov.Jointlcmm vcov.multlcmm fixef.hlme fixef.lcmm
+#' @aliases coef.hlme coef.lcmm coef.Jointlcmm coef.multlcmm coef.mpjlcmm
+#' coef.externSurv coef.externX vcov.hlme vcov.lcmm vcov.Jointlcmm vcov.multlcmm
+#' vcov.mpjlcmm vcov.externSurv vcov.externX fixef.hlme fixef.lcmm
 #' fixef.Jointlcmm fixef.multlcmm ranef.hlme ranef.lcmm ranef.Jointlcmm
 #' ranef.multlcmm fitted.hlme fitted.lcmm fitted.Jointlcmm fitted.multlcmm
 #' residuals.hlme residuals.lcmm residuals.Jointlcmm residuals.multlcmm
@@ -72,6 +74,16 @@ coef.Jointlcmm <- function(object,...)
 coef.mpjlcmm <- function(object,...)
     {
         estimates.mpjlcmm(object)
+}
+#' @export
+coef.externX <- function(object,...)
+    {
+        estimates.externX(object)
+    }
+#' @export
+coef.externSurv <- function(object,...)
+    {
+        estimates.externSurv(object)
     }
 
 
@@ -98,6 +110,16 @@ vcov.Jointlcmm <- function(object,...)
     }
 #' @export
 vcov.mpjlcmm <- function(object,...)
+    {
+        VarCov(object)
+}
+#' @export
+vcov.externX <- function(object,...)
+    {
+        VarCov(object)
+    }
+#' @export
+vcov.externSurv <- function(object,...)
     {
         VarCov(object)
     }