+++ inst/doc/lcmm.R
@@ -6,7 +6,7 @@ knitr::opts_chunk$set(
   
 )
 
-## ---- eval=FALSE--------------------------------------------------------------
+## ----eval=FALSE---------------------------------------------------------------
 #  hlme(fixed, mixture, random, subject, classmb, ng = 1, idiag = FALSE, nwg = FALSE, cor = NULL, data, B, convB = 0.0001, convL = 0.0001, convG = 0.0001, prior, maxiter = 500, subset = NULL, na.action = 1, posfix = NULL)
 
 ## ----comment=''---------------------------------------------------------------
@@ -15,7 +15,7 @@ head(paquid)
 ## ----comment=''---------------------------------------------------------------
 summary(paquid)
 
-## ---- results='hide',message=FALSE,warning=FALSE, echo=FALSE------------------
+## ----results='hide',message=FALSE,warning=FALSE, echo=FALSE-------------------
 library(lcmm)
 
 ## ----comment='', results='hide'-----------------------------------------------